refactor analysis tests
